+++ src/check.tcl
@@ -270,11 +270,12 @@ proc check::_create {} {
     {fra.radA - - - - {pack -side left -padx 9}  {-t "current file" -var ::alited::check::what -value 1}}
     {fra.radB - - - - {pack -side left -padx 9}  {-t "all of session files" -var ::alited::check::what -value 2}}
     {fra2 fra T 1 1 {-st nsew -pady 3 -padx 3} {-padding {5 5 5 5} -relief groove}}
-    {.butHelp - - - - {pack -side left} {-t "Help" -command ::alited::check::Help}}
+    {.ButHelp - - - - {pack -side left} {-t "Help" -command ::alited::check::Help}}
     {.h_ - - - - {pack -side left -expand 1 -fill both}}
     {.ButOK - - - - {pack -side left -padx 2} {-t "Check" -command ::alited::check::Ok}}
     {.butCancel - - - - {pack -side left} {-t Cancel -command ::alited::check::Cancel}}
   }
+  bind $win <F1> "[$obCHK ButHelp] invoke"
   if {[set geo $al(checkgeo)] ne {}} {
     set geo [string range $geo [string first + $geo] end]
     set geo "-geometry $geo"